Support and testing for Secure MongoDB clusters (CORE-789)
Note that secure MongoDB clusters were already supported, this PR simply
makes that more explicit by adding the ability to spin up a MongoDB
cluster with authentication for testing.

Using this it then adds a bunch of unit and integration tests around
connecting/failing to connect to the cluster with different combinations
of good/bad credentials.