Introduce LabelStores API (CORE-764)
Adds a new LabelStores API for mapping arbitrary keys to Label IDs.
This interface extends DictionaryLabelStore since it relies on the
dictionary for assigning IDs to labels.

Starting building out the early stages of a contract test suite for the
new interface and upgraded most existing implementations to fulfill the
new interface contract.